Rename SFTP settings.py to default_settings.py
