+++ tardis/default_settings/__init__.py
@@ -30,7 +30,7 @@ from .uploads import *
 from .urls import *
 
 # Default enabled app settings
-from ..apps.sftp.settings import *
+from ..apps.sftp.default_settings import *
 
 
 # Get version from git to be displayed on About page.
